fix: code refactor for latest bundle
